% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trasnlit.R
\name{translit}
\alias{translit}
\title{Transliterate string in Ukrainian}
\usage{
translit(string)
}
\arguments{
\item{string}{A string in Ukrainian}
}
\value{
A string in Latin
}
\description{
This function converts a given string from Ukrainian Cyrillic to
Latin characters using a specific set of transliteration rules.
}
